Feature/follow

+++ src/Controller/CommentController.php
@@ -3,19 +3,29 @@
 namespace App\Controller;
 
 use App\Entity\Comment;
+use App\Repository\SignalementRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class CommentController extends AbstractController
 {
-    #[Route('/{id}/delete', name: 'delete_comment')]
-    public function delete(Comment $comment): Response{
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($comment);
-        $entityManager->flush();
-        return $this->redirectToRoute('recipe_show', [
-            'id' => $comment->getRecette()->getId(),
-        ]);
+    #[Route('/delete/{id}', name: 'delete_comment')]
+    public function delete(Comment $comment, SignalementRepository $signalementRepository): Response{
+        $user = $this->getUser();
+        if(in_array("ROLE_ADMIN", $user->getRoles()) XOR $user->getId() == $comment->getUser()->getId()){
+            $signalement = $signalementRepository->findOneBy(['message' => $comment]);
+            $entityManager = $this->getDoctrine()->getManager();
+            if($signalement){
+                $entityManager->remove($signalement);
+            }
+            $entityManager->remove($comment);
+            $entityManager->flush();
+            return $this->redirectToRoute('recipe_show', [
+                'id' => $comment->getRecette()->getId(),
+            ]);
+        }else{
+            return $this->redirectToRoute('login');
+        }
     }
 }
